+++ src/Interfaces/gift/Gift.ts
@@ -49,12 +49,15 @@ export default interface Gift {
     shippingZip?: string;
     resourceType?: string;
     dispatchBy?: Date | string;
+    assigneeId?: number;
+    reference?: string;
     // relations
     giftBasket?: GiftBasket | null;
     order?: Order | null;
     user?: User | null;
     redeemedBy?: User | null;
     shippingAddons?: Array<ShippingAddon>;
+    assignee?: User | null;
     // accessors
     statusId?: number;
     deliveryMethod?: string;